[6CORE-023] moved warehouse dto to warehouse package.
