+++ 3-platform/warehouse/spec/src/main/java/com/_6core/platform/warehousespec/rest/v1/dto/warehouse/WarehouseRequest.java
@@ -1,4 +1,4 @@
-package com._6core.platform.warehousespec.rest.v1.dto;
+package com._6core.platform.warehousespec.rest.v1.dto.warehouse;
 
 import java.math.BigDecimal;
